requirements.txt
# requirements.txt
langchain>=0.1.0
chromadb>=0.4.15
transformers>=4.36.2
torch>=2.2.2
python-telegram-bot==20.7
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
numpy>=1.26.2
pydantic>=2.5.2